Fix #31 where error is received if a device has no sync group
- Having no device_sync_group_id in devices is accepted if the device
  was never added into a sync_group.

This bug occured becaused we didn't handle for a null
device_sync_group_id as it was mistaken that sql.ErrNoRows will trigger.

This is incorrect, instead we needed to make sure that the scanned
variable can be null (with a sql.NullInt64) and then handle it
accordingly when its null.

In this current codebase, if it is null, the returned value will still
be nil so there's no additional conditional check we need to do as the
output is still valid.